- added nice css classes for styling - cleaned up the responses of jsp files a little - first attempt at getting the ordering right.

+++ web/gallery.jsp
@@ -102,7 +102,11 @@ Author : maartenl
 
             function deleteGallery()
             {
-                alert("deleteGallery");
+                var r=confirm("Are you sure?");
+                if (r!=true)
+                {
+                    return;
+                }
                 var id = YourPersonalPhotographOrganiserBag.galleries[YourPersonalPhotographOrganiserBag.index].id;
                 $.ajax({
                     type: "DELETE",
